requirements.txt
absl-py==2.0.0
accelerate==0.22.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
av==11.0.0
backcall==0.2.0
beartype==0.16.4
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
cmake==3.27.6
comm==0.1.4
contourpy==1.1.1
cycler==0.12.0
datasets==2.14.6
debugpy==1.8.0
decorator==5.1.1
diffusers==0.21.4
dill==0.3.7
docstring-parser==0.15
einops==0.7.0
ema-pytorch==0.3.0
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.12.4
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.6.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
importlib-resources==6.1.0
inquirerpy==0.3.4
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
kornia==0.7.0
lit==17.0.1
llvmlite==0.41.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.1
numba==0.58.0
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.108
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pfzy==0.3.4
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pyDataverse==0.3.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.33.0
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.17.1
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.10.1
seaborn==0.13.2
simple-parsing==0.1.4
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.14.0
torch==2.1.0+cu118
torch-tb-profiler==0.4.3
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.34.0
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
